pubsub_to_fs_cf/requirements.txt
# Function dependencies, for example:
# package>=version
requests==2.24.0
python-telegram-bot==13.0
firebase-admin==4.4.0
google-cloud-bigquery==2.2.0
google-cloud-firestore==1.9.0